+++ mem-writer/ds_test.go
@@ -14,21 +14,21 @@ func init() {
 func TestWriterMap(t *testing.T) {
 	s := sugar.New(t)
 
-	s.Assert("`add` task to map", func(log sugar.Log) bool {
+	s.Assert("`add` task to map", func(logf sugar.Log) bool {
 		writerMap, err := psmbtcp.WriterDataStoreCreator("Writer")
 
-		log(err)
+		logf(err)
 		if err != nil {
 			return false
 		}
 
 		writerMap.Add("123456", "12")
-		log("add `123456` to table")
+		logf("add `123456` to table")
 		writerMap.Add("234561", "34")
-		log("add `234561` to table")
+		logf("add `234561` to table")
 
 		r1, b1 := writerMap.Get("123456")
-		log("get `123456` from table")
+		logf("get `123456` from table")
 		if !b1 {
 			return false
 		}
@@ -37,19 +37,19 @@ func TestWriterMap(t *testing.T) {
 		}
 
 		_, b2 := writerMap.Get("1234567")
-		log("get `1234567` from table")
+		logf("get `1234567` from table")
 
 		if b2 {
 			return false
 		}
 
 		writerMap.Delete("123456")
-		log("delete `123456` from table")
+		logf("delete `123456` from table")
 		writerMap.Delete("1234567")
-		log("delete `1234567` from table")
+		logf("delete `1234567` from table")
 
 		_, b3 := writerMap.Get("123456")
-		log("get `123456` from table")
+		logf("get `123456` from table")
 		if b3 {
 			return false
 		}
