+++ types_test.go
@@ -10,10 +10,10 @@ func TestTypes(t *testing.T) {
 
 	s := sugar.New(t)
 
-	s.Assert("Test MarshalJSON", func(log sugar.Log) bool {
+	s.Assert("Test MarshalJSON", func(logf sugar.Log) bool {
 		var a JSONableByteSlice
 		if _, err := a.MarshalJSON(); err != nil {
-			log(err)
+			logf(err)
 		}
 		return true
 	})
